added interceptor, controller/service tests update, separated DB for each user

+++ src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { Repository, DeleteResult } from 'typeorm';
@@ -13,41 +13,64 @@ export class UsersService {
     private usersRepository: Repository<User>,
   ) {}
 
-  create(createUserDto: CreateUserDto) {
-    return this.usersRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto, creatorId: string) {
+    const newUser = this.usersRepository.create({
+      ...createUserDto,
+      createdById: creatorId
+    });
+    return this.usersRepository.save(newUser);
   }
 
-  findAll() {
-    return this.usersRepository.find();
+  findAllByCreator(creatorId: string) {
+    return this.usersRepository.find({
+      where: { createdById: creatorId },
+    });
   }
 
-  async findOne(id: string): Promise<User> {
-    const user = await this.usersRepository.findOne({ where: { id } });
+  async findOne(id: string, requestingUserId: string): Promise<User> {
+    const user = await this.usersRepository.findOne({ 
+      where: { id },
+      relations: ['createdBy'],
+    });
+    
     if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
     }
+
+    if (user.id !== requestingUserId && user.createdById !== requestingUserId) {
+      throw new ForbiddenException('Access denied');
+    }
+
     return user;
   }
-
-  async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
-  const user = await this.usersRepository.findOne({ where: { id } });
   
-  if (!user) {
-    throw new NotFoundException(`User with ID ${id} not found`);
-  }
-
-
-  const updatedUser = this.usersRepository.merge(user, updateUserDto);
-  return await this.usersRepository.save(updatedUser);
-  }
-
-  async delete(id: string): Promise<DeleteResult> {
-    const res = await this.usersRepository.delete(id);
-    if (res.affected === 0) {
+  async update(id: string, updateUserDto: UpdateUserDto, requestingUserId: string): Promise<User> {
+    const user = await this.usersRepository.findOne({ where: { id } });
+    
+    if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
     }
 
-    return res;
+    if (user.id !== requestingUserId && user.createdById !== requestingUserId) {
+      throw new ForbiddenException('Access denied');
+    }
+
+    const updatedUser = this.usersRepository.merge(user, updateUserDto);
+    return await this.usersRepository.save(updatedUser);
+  }
+
+  async delete(id: string, requestingUserId: string): Promise<DeleteResult> {
+    const user = await this.usersRepository.findOne({ where: { id } });
+    
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
+    if (user.createdById !== requestingUserId) {
+      throw new ForbiddenException('Access denied');
+    }
+
+    return await this.usersRepository.delete(id);
   }
 
   // auth0service
@@ -59,14 +82,16 @@ export class UsersService {
 
     if (!user) {
       user = this.usersRepository.create({
-        ...auth0Data
+        auth0ID: auth0Data.auth0ID,
+        email: auth0Data.email,
+        firstName: '',
+        lastName: '', 
+        isRegistrationCompleted: false,
       });
-    } else {
-      user.email = auth0Data.email;
-      user.auth0ID = auth0Data.auth0ID;
+      user = await this.usersRepository.save(user);
     }
 
-    return await this.usersRepository.save(user);
+    return user;
   }
 
   async findByAuth0Id(auth0ID: string): Promise<User | null> {
